requirements.txt
pygithub
boto3